refactor: 패키지 구조 정리
- domain 레이어 생성
  - model 패키지를 domain 레이어 내부에 위치
  - Repository 인터페이스를 domain 레이어의 repository 패키지 내부에 위치

- data 레이어 패키지 구조 정리
  - local database 패키지 생성하여 DB 관련 클래스 정리
  - repository 패키지 생성하여 구현체 클래스 정리
