+++ app/src/main/java/woowacourse/shopping/data/local/ShoppingDatabase.kt
@@ -1,4 +1,4 @@
-package woowacourse.shopping.data
+package woowacourse.shopping.data.local
 
 import androidx.room.Database
 import androidx.room.RoomDatabase
