+++ app/src/main/java/woowacourse/shopping/data/mapper/ProductMapper.kt
@@ -1,7 +1,7 @@
 package woowacourse.shopping.data.mapper
 
-import woowacourse.shopping.data.CartProductEntity
-import woowacourse.shopping.model.Product
+import woowacourse.shopping.data.local.CartProductEntity
+import woowacourse.shopping.domain.model.Product
 
 fun Product.toEntity(): CartProductEntity {
     return CartProductEntity(
