+++ app/src/main/java/woowacourse/shopping/data/repository/CartDefaultRepository.kt
@@ -1,6 +1,7 @@
-package woowacourse.shopping.data
+package woowacourse.shopping.data.repository
 
-import woowacourse.shopping.model.Product
+import woowacourse.shopping.domain.repository.CartRepository
+import woowacourse.shopping.domain.model.Product
 
 // TODO: Step2 - CartProductDao를 참조하도록 변경
 class CartDefaultRepository : CartRepository {
