+++ app/src/main/java/woowacourse/shopping/domain/repository/CartRepository.kt
@@ -1,6 +1,6 @@
-package woowacourse.shopping.data
+package woowacourse.shopping.domain.repository
 
-import woowacourse.shopping.model.Product
+import woowacourse.shopping.domain.model.Product
 
 interface CartRepository {
     fun addCartProduct(product: Product)
