+++ app/src/main/java/woowacourse/shopping/domain/repository/ProductRepository.kt
@@ -1,0 +1,7 @@
+package woowacourse.shopping.domain.repository
+
+import woowacourse.shopping.domain.model.Product
+
+interface ProductRepository {
+    fun getAllProducts(): List<Product>
+}
