+++ app/src/main/java/woowacourse/shopping/ui/MainViewModel.kt
@@ -3,11 +3,9 @@ package woowacourse.shopping.ui
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
-import woowacourse.shopping.data.CartDefaultRepository
-import woowacourse.shopping.data.CartRepository
-import woowacourse.shopping.data.ProductDefaultRepository
-import woowacourse.shopping.data.ProductRepository
-import woowacourse.shopping.model.Product
+import woowacourse.shopping.domain.model.Product
+import woowacourse.shopping.domain.repository.CartRepository
+import woowacourse.shopping.domain.repository.ProductRepository
 
 class MainViewModel(
     private val productDefaultRepository: ProductRepository,
