+++ app/src/main/java/woowacourse/shopping/ui/cart/CartViewModelFactory.kt
@@ -2,7 +2,7 @@ package woowacourse.shopping.ui.cart
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelProvider
-import woowacourse.shopping.data.CartDefaultRepository
+import woowacourse.shopping.data.repository.CartDefaultRepository
 
 class CartViewModelFactory : ViewModelProvider.Factory {
     override fun <T : ViewModel> create(modelClass: Class<T>): T {
