+++ app/src/main/res/layout/item_product.xml
@@ -7,7 +7,7 @@
 
         <variable
             name="item"
-            type="woowacourse.shopping.model.Product" />
+            type="woowacourse.shopping.domain.model.Product" />
     </data>
 
     <androidx.constraintlayout.widget.ConstraintLayout
